+++ Package.swift
@@ -39,11 +39,13 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-nio-ssl.git", from: "2.0.0"),
         .package(url: "https://github.com/apple/swift-log.git", from: "1.0.0"),
         .package(url: "https://github.com/apple/swift-metrics.git", "1.0.0"..<"3.0.0"),
+        .package(url: "https://github.com/swift-server/async-http-client.git", from: "1.0.0")
     ],
     targets: [
         .target(
             name: "SmokeHTTPClient",
-            dependencies: ["Logging", "Metrics", "NIO", "NIOHTTP1", "NIOFoundationCompat", "NIOSSL"]),
+            dependencies: ["Logging", "Metrics", "NIO", "NIOHTTP1",
+                           "NIOFoundationCompat", "NIOSSL", "AsyncHTTPClient"]),
         .target(
             name: "QueryCoding",
             dependencies: ["ShapeCoding"]),
