+++ Sources/SmokeHTTPClient/HTTPOperationsClient +executeSyncWithOutput.swift
@@ -11,7 +11,7 @@
 // express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
 //
-//  HTTPClient+executeSyncWithOutput.swift
+//  HTTPOperationsClient+executeSyncWithOutput.swift
 //  SmokeHTTPClient
 //
 
@@ -22,7 +22,7 @@ import NIOSSL
 import NIOTLS
 import Logging
 
-public extension HTTPClient {
+public extension HTTPOperationsClient {
     /**
      Submits a request that will return a response body to this client synchronously.
      
@@ -34,7 +34,7 @@ public extension HTTPClient {
          - Returns: the response body.
          - Throws: If an error occurred during the request.
      */
-    func executeSyncWithOutput<InputType, OutputType, InvocationReportingType: HTTPClientInvocationReporting, HandlerDelegateType: HTTPClientChannelInboundHandlerDelegate>(
+    func executeSyncWithOutput<InputType, OutputType, InvocationReportingType: HTTPClientInvocationReporting, HandlerDelegateType: HTTPClientInvocationDelegate>(
         endpointOverride: URL? = nil,
         endpointPath: String,
         httpMethod: HTTPMethod,
@@ -64,7 +64,7 @@ public extension HTTPClient {
          - Throws: If an error occurred during the request.
      */
     internal func executeSyncWithOutputWithWrappedInvocationContext<InputType, OutputType, InvocationReportingType: HTTPClientInvocationReporting,
-            HandlerDelegateType: HTTPClientChannelInboundHandlerDelegate>(
+            HandlerDelegateType: HTTPClientInvocationDelegate>(
         endpointOverride: URL? = nil,
         endpointPath: String,
         httpMethod: HTTPMethod,
@@ -81,7 +81,7 @@ public extension HTTPClient {
                 completedSemaphore.signal()
             }
             
-            let channelFuture = try executeAsyncWithOutput(
+            _ = try executeAsyncWithOutput(
                 endpointOverride: endpointOverride,
                 endpointPath: endpointPath,
                 httpMethod: httpMethod,
@@ -90,23 +90,6 @@ public extension HTTPClient {
                 // the completion handler can be safely executed on a SwiftNIO thread
                 asyncResponseInvocationStrategy: SameThreadAsyncResponseInvocationStrategy<Result<OutputType, HTTPClientError>>(),
                 invocationContext: invocationContext)
-            
-            channelFuture.whenComplete { result in
-                switch result {
-                case .success(let channel):
-                    channel.closeFuture.whenComplete { _ in
-                        // if this channel is being closed and no response has been recorded
-                        if responseResult == nil {
-                            responseResult = .failure(HTTPClient.unexpectedClosureType)
-                            completedSemaphore.signal()
-                        }
-                    }
-                case .failure(let error):
-                    // there was an issue creating the channel
-                    responseResult = .failure(HTTPClientError(responseCode: 500, cause: error))
-                    completedSemaphore.signal()
-                }
-            }
             
             let logger = invocationContext.reporting.logger
             logger.debug("Waiting for response from \(endpointOverride?.host ?? endpointHostName) ...")
